pyproject.toml
[tool.poetry]
name = "arwa"
version = "0.1.2"
description = ""
authors = ["Abhinav Tushar <abhinav@lepisma.xyz>"]

[tool.poetry.dependencies]
python = "^3.8"
slackclient = "^2.8.2"
pydash = "^4.8.0"
gspread = "^3.6.0"
gcsa = "^0.3.2"
docopt = "^0.6.2"
jsonlines = "^1.2.0"
tqdm = "^4.49.0"
pytz = "^2020.4"
icalendar = "^4.0.7"
tabulate = "^0.8.7"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"

[tool.poetry.scripts]
arwa = "arwa.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
